requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi # for swagger implementation
uvicorn
sse_starlette